+++ csharp/src/Tsonic.CSharp.Node/fs/descriptor/readSync.cs
@@ -13,7 +13,7 @@ public static partial class fs
     /// <param name="length">The number of bytes to read.</param>
     /// <param name="position">The position in the file to start reading from. If null, reads from current position.</param>
     /// <returns>The number of bytes read.</returns>
-    public static int readSync(int fd, byte[] buffer, int offset, int length, int? position)
+    public static int readSync(int fd, byte[] buffer, int offset, int length, long? position)
     {
         if (buffer == null)
             throw new ArgumentNullException(nameof(buffer));
@@ -21,7 +21,7 @@ public static partial class fs
         return ReadSync(fd, buffer.AsSpan(), offset, length, position);
     }
 
-    private static int ReadSync(int fd, Span<byte> buffer, int offset, int length, int? position)
+    private static int ReadSync(int fd, Span<byte> buffer, int offset, int length, long? position)
     {
         var stream = FileDescriptorManager.Get(fd);
         if (stream == null)
@@ -51,7 +51,7 @@ public static partial class fs
     /// <param name="length">The number of bytes to read.</param>
     /// <param name="position">The position in the file to start reading from. If null, reads from current position.</param>
     /// <returns>The number of bytes read.</returns>
-    public static int readSync(int fd, Buffer buffer, int offset, int length, int? position)
+    public static int readSync(int fd, Buffer buffer, int offset, int length, long? position)
     {
         if (buffer == null)
             throw new ArgumentNullException(nameof(buffer));
